Clean up some of the duplicate testing functionality
